Remove check if image exists on filesystem
This removes the check if a image path stored in the DB does actually
point to a valid image file on the filesystem, as this emerged to be a
major performance problem.
The image path in the DB should only be set if the image was
successfully uploaded. On the dev,stage and prod servers every image
stored in the DB is also present on the filesystem. The following
command was used to check the images:
```for image in `sudo -u django psql -c "SELECT idea_image FROM
advocate_europe_ideas_idea WHERE idea_image != '';"`; do if [ ! -e
/home/django/django/media/$image ]; then echo $image; fi; done```

If the idea_image_exists is still needed it is advised to use
`os.path.exists` should be used instead of checking the width, as
checking the width includes opening and analyzing the file via PIL